backend/Server/requirements.txt
google-genai
requests
pillow
python-dotenv
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
transformers==4.35.0
torch==2.1.0
googletrans==4.0.0rc1
sentencepiece==0.1.99
accelerate==0.24.1
